requirements.txt
Django
psycopg2
oauthlib>=0.4.2
requests>=1.2.0
requests-oauthlib
django-tastypie
